extract gradietn radius calculations
