+++ src/shapes/index.ts
@@ -7,6 +7,7 @@ import parseEllipse from './parseEllipse'
 import * as Textures from 'textures'
 import { getBoundingBox } from './boundingBox'
 import { isStraightHandle } from './utils'
+import * as radialGradient from './radialGradient'
 
 // Internal collection for raw defs from <defs>
 type DefStop = { offset: number; color: [number, number, number, number]; opacity: number }
@@ -233,12 +234,11 @@ function toRuntimeGradient(resolved: AnyDef, boundingBox: BoundingBox): ShapePro
     const bbWidth = boundingBox.max_x - boundingBox.min_x
     const bbHeight = boundingBox.max_y - boundingBox.min_y
 
-    // 1) Absolute positions after gradientTransform
     const cAbs = applyLinearTransform(cx, cy, tf)
     const exAbs = applyLinearTransform(cx + r, cy, tf) // +X
     const eyAbs = applyLinearTransform(cx, cy + r, tf) // +Y
 
-    // 2) Normalize into shape's bbox uv space (same space the shader uses for uv/center/destination)
+    // Normalize positions
     const toNorm = (p: Point): Point => ({
       x: (p.x - boundingBox.min_x) / Math.max(bbWidth, 1e-8),
       y: 1 - (p.y - boundingBox.min_y) / Math.max(bbHeight, 1e-8),
@@ -247,76 +247,16 @@ function toRuntimeGradient(resolved: AnyDef, boundingBox: BoundingBox): ShapePro
     const ex = toNorm(exAbs)
     const ey = toNorm(eyAbs)
 
-    // 3) Build 2x2 linear map columns in normalized space
     const vx = { x: ex.x - center.x, y: ex.y - center.y }
     const vy = { x: ey.x - center.x, y: ey.y - center.y }
-
-    // If degenerate, fall back to simple case
-    const lenx = Math.hypot(vx.x, vx.y)
-    const leny = Math.hypot(vy.x, vy.y)
-    if (lenx < 1e-8 || leny < 1e-8) {
-      const horizontal_radius = Math.max(lenx, 1e-8)
-      const radius_ratio = lenx > 1e-8 ? leny / lenx : 1
-      const destination = {
-        x: center.x + (vx.x || 1) * (horizontal_radius / (lenx || 1)),
-        y: center.y + (vx.y || 0) * (horizontal_radius / (lenx || 1)),
-      }
-      return {
-        radial: {
-          stops,
-          center,
-          destination,
-          radius_ratio: isFinite(radius_ratio) ? radius_ratio : 1,
-        },
-      }
-    }
-
-    // 4) Principal axes via eigen-decomposition of A*A^T where A=[vx vy]
-    const Sxx = vx.x * vx.x + vy.x * vy.x
-    const Sxy = vx.x * vx.y + vy.x * vy.y
-    const Syy = vx.y * vx.y + vy.y * vy.y
-
-    const trace = Sxx + Syy
-    const det = Sxx * Syy - Sxy * Sxy
-    const tmp = Math.sqrt(Math.max(0, trace * trace * 0.25 - det))
-
-    // Largest then smallest eigenvalue
-    const lambda1 = trace * 0.5 + tmp
-    const lambda2 = trace * 0.5 - tmp
-
-    const s1 = Math.sqrt(Math.max(lambda1, 0))
-    const s2 = Math.sqrt(Math.max(lambda2, 0))
-
-    // Eigenvector for lambda1 (principal axis)
-    let ux = 0
-    let uy = 0
-    if (Math.abs(Sxy) > 1e-8 || Math.abs(lambda1 - Sxx) > 1e-8) {
-      // Solve (S - lambda1 I) u = 0; choose a stable component
-      if (Math.abs(Sxy) > Math.abs(lambda1 - Sxx)) {
-        ux = 1
-        uy = (lambda1 - Sxx) / (Sxy || 1e-12)
-      } else {
-        ux = (Sxy || 1e-12) / (lambda1 - Sxx || 1e-12)
-        uy = 1
-      }
-    } else {
-      ux = 1
-      uy = 0
-    }
-    const un = Math.hypot(ux, uy) || 1
-    ux /= un
-    uy /= un
-
-    // 5) Compose outputs expected by shader
-    const destination = { x: center.x + ux * s1, y: center.y + uy * s1 }
-    const radius_ratio = s1 > 1e-8 ? s2 / s1 : 1
+    const { destination, ratio } = radialGradient.getCorrectDestinationRatio(center, vx, vy)
 
     return {
       radial: {
         stops,
         center,
         destination,
-        radius_ratio: isFinite(radius_ratio) ? radius_ratio : 1,
+        radius_ratio: ratio,
       },
     }
   }
@@ -544,19 +484,6 @@ export function createShapes(
             const def = defs[m[1]]
             if (def) {
               const resolved = resolveRef(defs, def)
-
-              /*
-                  eyelid_right
-                  shape width: 41.9, height: 29.8
-
-                  start: x:29.5, y: 24.2
-                  end: x: 10.8, y: 13.6
-                */
-              if (props.id === 'eyelid_left') {
-                console.log(paths)
-                console.log(boundingBox)
-                // debugger
-              }
               const grad = toRuntimeGradient(resolved, boundingBox)
               if (grad) serializedProps.fill = grad
             }
